feat(auth): add logout endpoint and form request

+++ app/Http/Controllers/AuthController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\Auth\LoginRequest;
 use App\Http\Requests\Auth\RegisterRequest;
 use App\Models\User;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Validation\ValidationException;
 
@@ -12,6 +13,8 @@ class AuthController extends Controller
 {
   public function register(RegisterRequest $request)
   {
+    if ($request->user()) return response()->json(['message' => 'Already authenticated'], 409);
+
     $user = User::create($request->only('name', 'email', 'phone', 'password'));
 
     $token = $user->createToken('auth_token')->plainTextToken;
@@ -23,12 +26,15 @@ class AuthController extends Controller
     ], 201);
   }
 
-  public function login(LoginRequest $request) {
+  public function login(LoginRequest $request)
+  {
+    if ($request->user()) return response()->json(['message' => 'Already authenticated'], 409);
+
     $validated = $request->validated();
 
     $user = User::where('email', $validated['email'])->first();
 
-    if(! $user || ! Hash::check($validated['password'], $user->password)) throw ValidationException::withMessages(['email' => 'sorry these credentials does not match']);
+    if (! $user || ! Hash::check($validated['password'], $user->password)) throw ValidationException::withMessages(['email' => 'sorry these credentials does not match']);
 
     $user->tokens()->delete();
     $token = $user->createToken('auth_token')->plainTextToken;
@@ -38,5 +44,14 @@ class AuthController extends Controller
       'token' => $token,
       'token_type' => 'Bearer'
     ]);
+  }
+
+  public function logout(Request $request)
+  {
+    $user = $request->user();
+
+    $user->tokens()->delete();
+
+    return response()->noContent();
   }
 }
